add item to shopping cart (amount is double) use case

+++ src/java/main/serviceLayer/IService.java
@@ -113,7 +113,7 @@ public interface IService {
      * @param visitorName
      * @return
      */
-    public Response addItemToShoppingCart(ItemFacade itemToInsert, int amount, String shopName,
+    public Response addItemToShoppingCart(ItemFacade itemToInsert, double amount, String shopName,
                                           String visitorName);
 
     /**
